Commit adding a new algorithm for drawing reactions that can (ideally) be independent of the number of reactions.
NEW
---
 * New algorithm that is based on a hybrid drawing method. Reaction rates are distributed among a pool of groups. Each group is defined by a minimal and a maximal value that are successive powers of 2. Mathematically this means that when using a rejection method within one of these groups ensures that the probability that a reaction be rejected is less than 1/2. A parent class updates these groups and uses a rejection-free algorithm to determine in which group the next reaction will happen.
       * ratetoken.h: New class used for communication between `RateGroup` and `HybridRateContainer`. As the structure is very dynamic, `RateToken` bears all the information associated with a reaction rates at a fixed memory position, so that the container and the groups can happily move the pointers to token in memory, yielding minimal memory rewriting. Importantly, the token also bear the information about the rate's current location (as it keeps moving). One part is relevant to the container (what group the rate is in) and another is relevant to the group (position within group, current rate). A PassKey (based on two classes `GroupKey` and `RateKey`) pattern is used to make sure that only the appropriate partner can modifiy the elements that are relevant to him.
       * rategroup.h: New class storing `RateToken` whose rate is contained between a minimal and maximal value defined at construction. Token can be inserted and removed. A random token can be drawn according to the rate distribution using a rejection method.
       * hybridratecontainer.h/cpp: New class that creates the `RateGroup` to store `RateToken`. It starts with a single group and creates new ones when necessary. Each time a reaction rate needs to be updated, its token is taken away from the its current group and put into its new group. It draws a random reaction by drawing a random group according to the associated rates and then asks the chosen group to draw a reaction within the group.

 * tests/rategroup_test.cpp: new unit testing for `RateGroup`.
 * tests/hybridratecontainer_test.cpp: new unit testing for `HybridRateContainer`.
 * tests/file_test.cpp.template: new template file for unit testing.

MAJOR
-----
 * ratecontainer.h: removed the `find()` method from the interface as it was not necessary (not used by any client) and made little sense for the new container. The `random_index()` method was made virtual pure as it initially relied on the `find` method.
         * ratevector.h, ratetree.h: documentation of the `find()` method was moved to the child classes. `random_index()` was also moved to these classes as inlined members.

 * ratemanager.cpp/h: now uses the new `HybridRateContainer`.

 * tests/experimentalcumulative.h: made the class autonomous by adding appropriate `includes`. Added an output operator and const correctness. Moved the distance functions typically used with the class (from `randomhandler_test.cpp`).
        * tests/randomhandler_test.cpp: removed functions now included in experimentalcumulative.h (see above).

 * tests/ratetree_test.cpp: added a drawing statistics test.

MINOR
-----
 * solver.h: type of number of reactions was changed to `long long int` as the number of reactions is rapidly way larger than a billion.

 * main.cpp: added time measurement.

 * ratetree.h: removed `virtual` keyword in members.

 * file.h.template: added a `project includes` section.

+++ src/solver.h
@@ -89,7 +89,7 @@ class Solver
    * @brief Accessor to the number of reactions that have occurred.
    * @return Number of reactions that have occurred from the beginning of simulation.
    */
-  int number_reactions_performed (void) const;  
+  long long int number_reactions_performed (void) const;  
   
 
   // ==========================
@@ -162,7 +162,7 @@ class Solver
   double _t;
 
   /** @brief Number of reactions that have been performed. */
-  int _number_reactions_performed;
+  long long int _number_reactions_performed;
 
   // =================
   //  Private Methods
@@ -179,7 +179,7 @@ inline double Solver::time (void) const
   return _t;
 }
 
-inline int Solver::number_reactions_performed (void) const 
+inline long long int Solver::number_reactions_performed (void) const 
 {
   return _number_reactions_performed;
 }
